requirements.txt
flask==2.3.3
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
spacy==3.7.2
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1